+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Added
 - Added TLS certificate expiry event: Notifications for upcoming LND TLS certificate expiration with configurable threshold before notification. (@Primexz)
+- Notification batching: Support for batching notifications to reduce frequency while ensuring important events are delivered promptly. Configurable flush interval and maximum batch size. (@Primexz)
 
 ### Deprecated
 ### Fixed
@@ -22,6 +23,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Prevented multiple notifications from being sent when lndClient.SubscribeEvents fails (e.g., during lnd startup when server is not active) and lndnotify is retried externally. (@Feelancer21)
 - Fixed issue where channel up events and chain sync lost events were incorrectly triggered. (@Primexz)
 
+### Changed
+
+- Updated [shoutrrr](https://github.com/nicholas-fedor/shoutrrr/releases/tag/v0.10.3)
+- Updated Golang to version 1.25.3 (@Primexz)
 
 ## [1.6.2] - 2025-10-18